Create N3h-kr3h-mah
This is my creation/invention/idea/plan from old dollar bills to new dollar bills being printed on plastic with magnetic strips bound inside of the plastic bound material for easier/cleaner scanning, tracking and distributing of dollar bills! Wondering how to protect the moving of money this way similar to ink dyes in bags if there was to be a heist!?! Mini motorized razor blade propellers on the inside of the money bags with a key to activate/deactivate the mechanism upon opening and closing the bags with plastic dollar bills. The security does the same jobs as before but with a new device to protect upon ant-theft. The propellers inside bag would activate inside shredding the plastic money without a key/code to deactivate! This obviously gives us more work, jobs and things to do on a respectable note; people have there own positions, respect, and selfworth to obtain when were not all the same! we all like to have something to do and think into! People need shit do! This might need a recalling for all old bills and would have to be called in with a certain date to be traded in for the new plastic bills upon use of the new and improved dollar bills. N.W.O or just an addition to curency similar to bitcoin!?!......The money! Nor does it ever grow on trees but is needed to make the world go round!...... so with a perspective on whos this and whats that!?! Whats that and whos this!?! No, its like this! Just different perspectives on the stuff! But whats the objective here to begin with!?! Money and stuff whether right or wrong or just breaking it all down to the rough! No view, perspective, opinion is better or worse, more wrong or right than the other! Just different; differences that should and need to be kept that way along with the system; systems; systematic ways of doibg thibgs with or without technology. Anyways.... remembering the golden rule: reduce, reuse, recycle; along with your 7 p's (prior, proper, planning, prevents, piss, poor, production) makes things that much better either which way! Respectfully, Antonio Elias Henao